Bump flutter_quill version to at least 3.8.7
With the 3.5.0 and 3.8.6 update of flutter_quill, the exports were
changed so that library internal files can no longer be imported
easily. The imports have now been fixed in this library, but this
makes delta_markdown incompatible with versions prior to 3.8.7.
Therefore the minimum version of flutter_quill was raised to 3.8.7.